requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
uuid==1.30
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==0.30.0
flask-migrate==4.1.0
flask-jwt-extended==4.7.1
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
apispec==6.8.1
flask-apispec==0.11.4
